requirements.txt
# Function dependencies, for example:
# package>=version
urllib3
requests
google-cloud-pubsub
pymongo[srv]
flask
pytz